requirements.txt
PyQt5
opencv-python-headless
pytesseract
Pillow
numpy
